+++ resolwe_bio/kb/tests/test_mapping.py
@@ -80,6 +80,14 @@ class MappingTestCase(APITestCase, TestCase):
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
+    def test_nonexistent_field(self):
+        MAPPING_URL = reverse("resolwebio-api:kb_mapping_search")
+
+        # Test with wrong query parameter
+        response = self.client.get(MAPPING_URL, {"nonexistent": "field"}, format="json")
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertEqual(response.status_text, "Bad Request")
+
     def test_serialization(self):
         MAPPING_URL = reverse("resolwebio-api:kb_mapping_search")
         kwargs = {"source_id": "FT1"}